requirements.txt
torch
torchaudio
torchvision
hydra-core
accelerate
transformers
librosa
torchcrepe
pandas
scipy
numpy
tensorboard
soundfile
datasets
einops
tqdm
auraloss
safetensors
matplotlib 